Move create group view models to special folder.
