+++ src/StudentToolkit/MVVM/ViewModels/Presentation/Group/Create/AddStudentsToGroupViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using StudentToolkit.MVVM.ViewModels.Presentation.GroupInfo;
 
-namespace StudentToolkit.MVVM.ViewModels.Presentation.Group;
+namespace StudentToolkit.MVVM.ViewModels.Presentation.Group.Create;
 
 public class AddStudentsToGroupViewModel : ViewModel
 {
