+++ src/StudentToolkit/MVVM/ViewModels/Presentation/Group/Create/InputGroupDataViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using StudentToolkit.MVVM.ViewModels.Presentation.GroupInfo;
 using StudentToolkit.WpfCore.Common.Helpers;
 
-namespace StudentToolkit.MVVM.ViewModels.Presentation.Group;
+namespace StudentToolkit.MVVM.ViewModels.Presentation.Group.Create;
 
 public class InputGroupDataViewModel : ViewModel
 {
