+++ src/StudentToolkit/Properties/GlobalUsings.cs
@@ -28,7 +28,7 @@ global using StudentToolkit.MVVM.ViewModels;
 global using StudentToolkit.MVVM.ViewModels.Base;
 global using StudentToolkit.MVVM.ViewModels.Components;
 global using StudentToolkit.MVVM.ViewModels.Model;
-global using StudentToolkit.MVVM.ViewModels.Presentation.Group;
+global using StudentToolkit.MVVM.ViewModels.Presentation.Group.Create;
 global using StudentToolkit.MVVM.Views.Windows;
 global using StudentToolkit.WpfCore.Commands.Base;
 global using StudentToolkit.WpfCore.Commands.Navigation;
